[Refactor] Move shuttleDirections.ts under directions/ folder
